extract interface for http request factory (in preparation for providing v2 implementation)
